Complete New User Workflow including anonymous user. Some fixes to bugs that happened only when user had no milestones.

+++ ios/Milestones/Milestones/HistoryViewTableModel.m
@@ -71,7 +71,7 @@
 }
 
 
--(void) loadMilestonesPage:(int) startIndex forTimePeriod:(NSString*) timePeriod withBlock:(PFArrayResultBlock) block {
+-(void) loadMilestonesPage:(NSInteger) startIndex forTimePeriod:(NSString*) timePeriod withBlock:(PFArrayResultBlock) block {
   // TODO: caching!
   if(self.baby) {
     [PFCloud callFunctionInBackground:@"queryMyMilestones"
